fix #10: Correct event handler function(s)
@alor pointed out that the `onEdgesChange` function should utilize the
`applyEdgeChanges` function from ReactFlow, not just the `addEdge`
function. To maintain support for creating a new edge, I added a new
`onConnect` handler to work exclusively with the `addEdge` function,
preserving the initial behavior.

(This change does not affect the source code of useUndoable and thus
will not constitute a version bump or release)